+++ src/Model/GameObject/NetworkedItem.hpp
@@ -34,14 +34,14 @@ public:
 	 *
 	 * @param delta_time
 	 */
-	void update(float delta_time) override;
+	void update(double delta_time) override;
 
 	/**
 	 * @brief Updates the body object with fixed delta time.
 	 *
 	 * @param delta_time
 	 */
-	void fixed_update(float delta_time) override {}
+	void fixed_update(double delta_time) override {}
 
 	/**
 	 * @brief Adds the draw call to the renderer.
